wifi and comments
