+++ app/src/main/java/com/openpositioning/PositionMe/sensors/KFLinear2D.java
@@ -2,10 +2,10 @@ package com.openpositioning.PositionMe.sensors;
 
 public class KFLinear2D {
     // State = [ x, y, vx, vy ] in meters
-    private double[] x;      // 长度为4
-    private double[][] P;    // 4x4 协方差矩阵
-    private double[][] Q;    // 4x4 过程噪声
-    private double[][] R;    // 2x2 测量噪声
+    private double[] x;      // Length is 4
+    private double[][] P;    // 4x4 covariance matrix
+    private double[][] Q;    // 4x4 process noise
+    private double[][] R;    // 2x2 measurement noise
 
     public KFLinear2D(double[] initialState,
                       double[][] initialCov,
@@ -45,8 +45,8 @@ public class KFLinear2D {
 
     public void update(double[] z) {
         double[][] H = {
-                {1,0,0,0},
-                {0,1,0,0}
+                {1, 0, 0, 0},
+                {0, 1, 0, 0}
         };
         double[] zPred = multiplyMatrixVector(H, x);
         double[] y = subtractVectors(z, zPred);
@@ -72,19 +72,19 @@ public class KFLinear2D {
         return new double[]{ x[0], x[1] };
     }
 
-    // 新增方法：返回协方差矩阵 P 的副本
+    // New method: returns a copy of the covariance matrix P
     public double[][] getErrorCovariance() {
         return copyMatrix(P);
     }
 
-    // -------------- 辅助矩阵运算方法 ----------------
+    // -------------- Auxiliary matrix operations methods ----------------
 
     private double[][] addMatrices(double[][] A, double[][] B) {
         int rows = A.length;
         int cols = A[0].length;
         double[][] out = new double[rows][cols];
-        for (int r=0; r<rows; r++){
-            for (int c=0; c<cols; c++){
+        for (int r = 0; r < rows; r++) {
+            for (int c = 0; c < cols; c++) {
                 out[r][c] = A[r][c] + B[r][c];
             }
         }
@@ -93,7 +93,7 @@ public class KFLinear2D {
 
     private double[] addVectors(double[] a, double[] b) {
         double[] out = new double[a.length];
-        for (int i=0; i<a.length; i++){
+        for (int i = 0; i < a.length; i++) {
             out[i] = a[i] + b[i];
         }
         return out;
@@ -101,7 +101,7 @@ public class KFLinear2D {
 
     private double[] subtractVectors(double[] a, double[] b) {
         double[] out = new double[a.length];
-        for (int i=0; i<a.length; i++){
+        for (int i = 0; i < a.length; i++) {
             out[i] = a[i] - b[i];
         }
         return out;
@@ -111,8 +111,8 @@ public class KFLinear2D {
         int rows = A.length;
         int cols = A[0].length;
         double[][] out = new double[rows][cols];
-        for (int r=0; r<rows; r++){
-            for (int c=0; c<cols; c++){
+        for (int r = 0; r < rows; r++) {
+            for (int c = 0; c < cols; c++) {
                 out[r][c] = A[r][c] - B[r][c];
             }
         }
@@ -124,11 +124,11 @@ public class KFLinear2D {
         int cols = B[0].length;
         int common = A[0].length;
         double[][] out = new double[rows][cols];
-        for (int i=0; i<rows; i++){
-            for (int j=0; j<cols; j++){
-                double sum=0;
-                for (int k=0; k<common; k++){
-                    sum += A[i][k]*B[k][j];
+        for (int i = 0; i < rows; i++) {
+            for (int j = 0; j < cols; j++) {
+                double sum = 0;
+                for (int k = 0; k < common; k++) {
+                    sum += A[i][k] * B[k][j];
                 }
                 out[i][j] = sum;
             }
@@ -136,62 +136,62 @@ public class KFLinear2D {
         return out;
     }
 
-    private double[] multiplyMatrixVector(double[][] M, double[] v){
+    private double[] multiplyMatrixVector(double[][] M, double[] v) {
         int rows = M.length;
         int cols = M[0].length;
         double[] out = new double[rows];
-        for (int r=0; r<rows; r++){
-            double sum=0;
-            for (int c=0; c<cols; c++){
-                sum += M[r][c]*v[c];
+        for (int r = 0; r < rows; r++) {
+            double sum = 0;
+            for (int c = 0; c < cols; c++) {
+                sum += M[r][c] * v[c];
             }
             out[r] = sum;
         }
         return out;
     }
 
-    private double[][] transpose(double[][] A){
+    private double[][] transpose(double[][] A) {
         int rows = A.length;
         int cols = A[0].length;
         double[][] T = new double[cols][rows];
-        for (int r=0; r<rows; r++){
-            for (int c=0; c<cols; c++){
+        for (int r = 0; r < rows; r++) {
+            for (int c = 0; c < cols; c++) {
                 T[c][r] = A[r][c];
             }
         }
         return T;
     }
 
-    private double[][] identityMatrix(int n){
+    private double[][] identityMatrix(int n) {
         double[][] I = new double[n][n];
-        for (int i=0; i<n; i++){
+        for (int i = 0; i < n; i++) {
             I[i][i] = 1.0;
         }
         return I;
     }
 
-    private double[][] copyMatrix(double[][] src){
+    private double[][] copyMatrix(double[][] src) {
         double[][] dst = new double[src.length][src[0].length];
-        for (int i=0; i<src.length; i++){
+        for (int i = 0; i < src.length; i++) {
             System.arraycopy(src[i], 0, dst[i], 0, src[i].length);
         }
         return dst;
     }
 
-    private double[][] invert2x2(double[][] M){
-        if (M.length!=2 || M[0].length!=2){
+    private double[][] invert2x2(double[][] M) {
+        if (M.length != 2 || M[0].length != 2) {
             throw new IllegalArgumentException("invert2x2 only for 2x2 matrix.");
         }
-        double det = M[0][0]*M[1][1] - M[0][1]*M[1][0];
-        if (Math.abs(det)<1e-12){
+        double det = M[0][0] * M[1][1] - M[0][1] * M[1][0];
+        if (Math.abs(det) < 1e-12) {
             throw new RuntimeException("Matrix is singular or near singular.");
         }
-        double invDet = 1.0/det;
+        double invDet = 1.0 / det;
         double[][] out = new double[2][2];
-        out[0][0] =  M[1][1]*invDet;
-        out[0][1] = -M[0][1]*invDet;
-        out[1][0] = -M[1][0]*invDet;
-        out[1][1] =  M[0][0]*invDet;
+        out[0][0] = M[1][1] * invDet;
+        out[0][1] = -M[0][1] * invDet;
+        out[1][0] = -M[1][0] * invDet;
+        out[1][1] = M[0][0] * invDet;
         return out;
     }
 }
